+++ src/application-navigation/index.js
@@ -1,0 +1,10 @@
+import {
+  titleConfigPropType, navigationItemsPropType, extensionItemsPropType, utilityItemsPropType, userConfigPropType,
+} from 'terra-application-navigation/lib/utils/propTypes';
+import ApplicationNavigation from './ApplicationNavigation';
+
+export default ApplicationNavigation;
+export {
+  titleConfigPropType, navigationItemsPropType, extensionItemsPropType, utilityItemsPropType, userConfigPropType,
+};
+