+++ src/disclosure-manager/_DisclosureContainer.jsx
@@ -1,10 +1,13 @@
-import React, { useEffect, useContext, useRef } from 'react';
+import React, {
+  useEffect, useContext, useRef, useMemo,
+} from 'react';
 import PropTypes from 'prop-types';
+import { injectIntl } from 'react-intl';
 
 import DisclosureManagerContext from 'terra-disclosure-manager/lib/DisclosureManagerContext';
 
 import { ApplicationLoadingOverlayProvider } from '../application-loading-overlay';
-import { NavigationPromptCheckpoint, navigationPromptResolutionOptionsShape } from '../navigation-prompt';
+import { NavigationPromptCheckpoint, navigationPromptResolutionOptionsShape, getUnsavedChangesPromptOptions } from '../navigation-prompt';
 import ApplicationErrorBoundary from '../application-error-boundary';
 
 const propTypes = {
@@ -20,9 +23,16 @@ const propTypes = {
   navigationPromptResolutionOptions: navigationPromptResolutionOptionsShape,
 };
 
-const DisclosureContainer = ({ children, navigationPromptResolutionOptions }) => {
+/**
+ * injectIntl (rather that the ApplicationIntlContext) is used here to keep the DisclosureContainer's context dependencies
+ * passive with previous v1.x versions. ApplicationIntlContext cannot be used here without a version bump.
+ */
+
+const DisclosureContainer = injectIntl(({ intl, children, navigationPromptResolutionOptions }) => {
   const disclosureManager = useContext(DisclosureManagerContext);
   const promptCheckpointRef = useRef();
+
+  const defaultPromptOptions = useMemo(() => getUnsavedChangesPromptOptions(intl), [intl]);
 
   useEffect(() => {
     disclosureManager.registerDismissCheck(() => new Promise((resolve, reject) => {
@@ -31,9 +41,9 @@ const DisclosureContainer = ({ children, navigationPromptResolutionOptions }) =>
         return;
       }
 
-      promptCheckpointRef.current.resolvePrompts(navigationPromptResolutionOptions).then(resolve, reject);
+      promptCheckpointRef.current.resolvePrompts(navigationPromptResolutionOptions || defaultPromptOptions).then(resolve, reject);
     }));
-  }, [disclosureManager, navigationPromptResolutionOptions]);
+  }, [defaultPromptOptions, disclosureManager, navigationPromptResolutionOptions]);
 
   return (
     <ApplicationErrorBoundary>
@@ -46,7 +56,7 @@ const DisclosureContainer = ({ children, navigationPromptResolutionOptions }) =>
       </ApplicationLoadingOverlayProvider>
     </ApplicationErrorBoundary>
   );
-};
+});
 
 DisclosureContainer.propTypes = propTypes;