+++ src/navigation-prompt/index.js
@@ -1,4 +1,5 @@
 import PropTypes from 'prop-types';
+import getUnsavedChangesPromptOptions from './getUnsavedChangesPromptOptions';
 
 const navigationPromptResolutionOptionsShape = PropTypes.oneOfType([PropTypes.shape({
   /**
@@ -37,4 +38,4 @@ const navigationPromptResolutionOptionsShape = PropTypes.oneOfType([PropTypes.sh
 
 export { default } from 'terra-navigation-prompt';
 export { NavigationPromptCheckpoint, PromptRegistrationContext } from 'terra-navigation-prompt';
-export { navigationPromptResolutionOptionsShape };
+export { navigationPromptResolutionOptionsShape, getUnsavedChangesPromptOptions };